Migrate length tests
Adding in needed Vector -> Interval infra
