+++ src/webgpu/shader/execution/expression/call/builtin/length.spec.ts
@@ -10,15 +10,10 @@ Returns the length of e (e.g. abs(e) if T is a scalar, or sqrt(e[0]^2 + e[1]^2 +
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../../../gpu_test.js';
 import { TypeF32, TypeVec } from '../../../../../util/conversion.js';
-import { lengthInterval } from '../../../../../util/f32_interval.js';
+import { FP } from '../../../../../util/floating_point.js';
 import { fullF32Range, vectorF32Range } from '../../../../../util/math.js';
 import { makeCaseCache } from '../../case_cache.js';
-import {
-  allInputSources,
-  generateUnaryToF32IntervalCases,
-  generateVectorToF32IntervalCases,
-  run,
-} from '../../expression.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
 
@@ -26,25 +21,41 @@ export const g = makeTestGroup(GPUTest);
 
 export const d = makeCaseCache('length', {
   f32: () => {
-    return generateUnaryToF32IntervalCases(fullF32Range(), 'unfiltered', lengthInterval);
+    return FP.f32.generateScalarToIntervalCases(
+      fullF32Range(),
+      'unfiltered',
+      FP.f32.lengthInterval
+    );
   },
   f32_vec2_const: () => {
-    return generateVectorToF32IntervalCases(vectorF32Range(2), 'finite', lengthInterval);
+    return FP.f32.generateVectorToIntervalCases(vectorF32Range(2), 'finite', FP.f32.lengthInterval);
   },
   f32_vec2_non_const: () => {
-    return generateVectorToF32IntervalCases(vectorF32Range(2), 'unfiltered', lengthInterval);
+    return FP.f32.generateVectorToIntervalCases(
+      vectorF32Range(2),
+      'unfiltered',
+      FP.f32.lengthInterval
+    );
   },
   f32_vec3_const: () => {
-    return generateVectorToF32IntervalCases(vectorF32Range(3), 'finite', lengthInterval);
+    return FP.f32.generateVectorToIntervalCases(vectorF32Range(3), 'finite', FP.f32.lengthInterval);
   },
   f32_vec3_non_const: () => {
-    return generateVectorToF32IntervalCases(vectorF32Range(3), 'unfiltered', lengthInterval);
+    return FP.f32.generateVectorToIntervalCases(
+      vectorF32Range(3),
+      'unfiltered',
+      FP.f32.lengthInterval
+    );
   },
   f32_vec4_const: () => {
-    return generateVectorToF32IntervalCases(vectorF32Range(4), 'finite', lengthInterval);
+    return FP.f32.generateVectorToIntervalCases(vectorF32Range(4), 'finite', FP.f32.lengthInterval);
   },
   f32_vec4_non_const: () => {
-    return generateVectorToF32IntervalCases(vectorF32Range(4), 'unfiltered', lengthInterval);
+    return FP.f32.generateVectorToIntervalCases(
+      vectorF32Range(4),
+      'unfiltered',
+      FP.f32.lengthInterval
+    );
   },
 });
